requirements.txt
fastapi
uvicorn
requests
pdfplumber
python-multipart
gspread
google-auth